This is a project file

+++ main.h
@@ -12,7 +12,7 @@ int octal(const char *format, va_list ptr);
 int _putchar(char c);
 int _printf_aid(const char *format, va_list ptr);
 int _printf_aid_1(const char *format, va_list ptr);
-int print_integer(int n);
+int print_integer(long int n);
 int print_binary(unsigned int n);
 int _printf_aid_2(const char *format, va_list ptr);
 int refactor_printf(const char *format, va_list ptr);
